+++ scripts/update-homebrew.sh
@@ -109,12 +109,10 @@ class Kipferl < Formula
     end
 
     bin.install binary_name => "kipferl"
-    bin.install_symlink "kipferl" => "ucharm"
   end
 
   test do
     assert_match "Kipferl", shell_output("#{bin}/kipferl --version")
-    assert_match "renamed to", shell_output("#{bin}/ucharm --version 2>&1")
   end
 end
 EOF